# fix(perf): Various histogram fixes