--- a/static/app/views/performance/transactionSummary/latencyChart.tsx
+++ b/static/app/views/performance/transactionSummary/latencyChart.tsx
@@ -209,19 +209,7 @@ class LatencyChart extends React.Component<Props, State> {
       location
     );
 
-    let min: number | undefined = undefined;
-    let max: number | undefined = undefined;
-
-    if (ZOOM_START in location.query) {
-      min = decodeInteger(location.query[ZOOM_START], 0);
-    }
-
-    if (ZOOM_END in location.query) {
-      const decodedMax = decodeInteger(location.query[ZOOM_END]);
-      if (typeof decodedMax === 'number') {
-        max = decodedMax;
-      }
-    }
+    const {min, max} = decodeHistogramZoom(location);
 
     const field = filterToField(currentFilter) ?? 'transaction.duration';
 
@@ -293,6 +281,24 @@ export function LatencyChartControls(props: {location: Location}) {
       }}
     </Histogram>
   );
+}
+
+export function decodeHistogramZoom(location: Location) {
+  let min: number | undefined = undefined;
+  let max: number | undefined = undefined;
+
+  if (ZOOM_START in location.query) {
+    min = decodeInteger(location.query[ZOOM_START], 0);
+  }
+
+  if (ZOOM_END in location.query) {
+    const decodedMax = decodeInteger(location.query[ZOOM_END]);
+    if (typeof decodedMax === 'number') {
+      max = decodedMax;
+    }
+  }
+
+  return {min, max};
 }
 
 export default LatencyChart;