Histogram query strings (i.e. `startDuration` and `endDuration`) should implicitly be applied to various queries on the Performance Summary page. In other words, is the user selects a range of 200ms to 400ms range from the histogram, then `transaction.duration:>200ms transaction.duration:<400ms` are implicitly added to the search conditions. This also applies for span op breakdown filters.

https://user-images.githubusercontent.com/139499/115475364-6eac6b00-a20d-11eb-9d72-79c8b766a340.mp4

In addition, histogram query strings are removed when switching Display chart modes from the "Duration Distribution" chart.